new structure and logic

+++ src/migrations/m160905_111034_dotplant_structure_add_to_entity.php
@@ -10,11 +10,6 @@ class m160905_111034_dotplant_structure_add_to_entity extends Migration
     {
         $this->addColumn(
             Entity::tableName(),
-            'edit_route',
-            $this->string(255)->notNull()->defaultValue('')
-        );
-        $this->addColumn(
-            Entity::tableName(),
             'tree_icon',
             $this->string(255)->notNull()->defaultValue('')
         );
@@ -32,7 +27,10 @@ class m160905_111034_dotplant_structure_add_to_entity extends Migration
 
     public function down()
     {
-        $this->dropColumn(Entity::tableName(), 'edit_route');
+        $schema = $this->db->getTableSchema(Entity::tableName());
+        if (true === in_array('edit_route', $schema->columnNames)) {
+            $this->dropColumn(Entity::tableName(), 'edit_route');
+        }
         $this->dropColumn(Entity::tableName(), 'tree_icon');
         $this->dropColumn(StructureTranslation::tableName(), 'can_stop_url_parsing');
         $this->dropColumn(StructureTranslation::tableName(), 'url_processing_params');
